#53 Let people sign pre-hashed messages with ECDSA.
This feature won't be available for now when using the registry to avoid mistakes when signing with ECDSA under RFC 6979. Nevertheless, people would still be able to use an ECDSASigner directly and specify any values they want for the Digest (including null for no digest) and the MAC used for RFC 6967 (including null to avoid RFC 6979). See dartdoc of constructor for more information.