Dynamic Registry implementation
The implementation is based on Reflectable, a library that does efficient reflection in Dart. So, instead of registering all implemented algorithms manually, the implementation itself defines how it should be registered. The Registry class will then search through all imported cipher.* libraries to look for implementations to include.

This allows users to import api.dart and cherry pick certain implementations and still be able to use the generic constructors. It also makes it easier to add new implementations, adding a new file is all that it takes.

This also allows external libraries to add implementations to cipher, just by creating cipher.* libraries and importing them.

A side change is that I had to create a separate class for all ec curves. This positive side of this is again that users that only need one curve, don't have to import all curves.

Other changes: I moves the CTRStreamCipher and SICStreamCipher modes into the cipher.stream namespace and created CTRBlockCipher and SICBlockCipher using the StreamCipherAsBlockCipher class for completeness.